Do not constrain the max CPU frequency for apps on setting the string value.
It is already constrained on editing and when actually applying the frequency.
The string value constrain is problematic because it happens in a class that
has no knowledge of the actual CPU limits, so it can only contrain to
hardcoded limits.

With this change you'll actually be able to run apps at 432 MHz.