Removed ability to configure custom working directory.
Working directory was only configurable in the ini files, not in the GUI.

Many GP2X/Dingoo applications break if the working directory is changed
because they use relative paths for locating their data files.

I don't see a reason to keep this feature, since I don't know of a realistic
scenario where it would be useful. In any case, the functionality can be
re-created by using a wrapper script to change the working directory based on
a parameter value.